% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_datasets.R
\name{comp_datasets}
\alias{comp_datasets}
\alias{ggplot_comp_datasets}
\title{Compare Dataset}
\usage{
comp_datasets(
  traitModule,
  ds1 = datasets,
  ds2 = datasets[datasets != ds1],
  dataset = "Metab",
  response = "value",
  ...
)

ggplot_comp_datasets(
  object,
  ds = c("PlaMet0", "PlaMet120"),
  abs = FALSE,
  ...,
  compmods = comp_datasets(object, ds[1], ds[2], ...)
)
}
\arguments{
\item{traitModule}{module from `moduleHarmony`}

\item{ds1}{first dataset within `dataset`}

\item{ds2}{second dataset within `dataset`}

\item{dataset}{dataset in `traitModule`}

\item{response}{response in `dataset`}

\item{...}{additional arguments ignored}

\item{object}{from `moduleHarmony`}

\item{ds}{pair of datasets to plot}

\item{abs}{absolute kME if `TRUE`}
}
\value{
object of class `module_kMEs`

GGplot object
}
\description{
Compare Dataset

GGplot including creation of compare dataset object
}
