% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{harmonize}
\alias{harmonize}
\title{Harmonize Dataset}
\usage{
harmonize(
  dataset,
  links,
  userHarmony,
  ...,
  normalize = TRUE,
  condition_name = "condition",
  minstrains = nstrains - 2,
  drop_na_pvalue = TRUE
)
}
\arguments{
\item{dataset}{name of dataset}

\item{links}{data frame with `shortname`, `address`, `longname`}

\item{userHarmony}{function provided by user}

\item{...}{additional optional parameters to `userHarmony()`}

\item{normalize}{apply `normalscores` if `TRUE`}

\item{condition_name}{name of `condition` column if present.}

\item{minstrains}{minimum number of `strains` allowed per `trait`}

\item{drop_na_pvalue}{drop traits with any missing pvalues if `TRUE`; if `FALSE` return `traitStats` object only}
}
\value{
side action to save RDS files locally
}
\description{
Harmonize Dataset
}
