% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_dataset_modules.R
\name{join_dataset_modules}
\alias{join_dataset_modules}
\title{Join Dataset Modules}
\usage{
join_dataset_modules(dmods, response)
}
\arguments{
\item{dmods}{list of objects of class `listof_wgcnaModules`}

\item{response}{names of response to extract}
}
\value{
object of class `listof_wgcnaModules`
}
\description{
Join Dataset Modules
}
