% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkpath.R
\name{link_datasets}
\alias{link_datasets}
\title{Link datasets}
\usage{
link_datasets(object, linkfile, deployDir = "deploy")
}
\arguments{
\item{object}{data frame}

\item{linkfile}{CSV file with shortname, URL, longname}

\item{deployDir}{directory with `help.Rmd` for `help.md`}
}
\value{
invisible
}
\description{
Side effects to save RDS `datasets` data frame and `help.md` about file.
}
