% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkpath.R
\name{linkpath}
\alias{linkpath}
\title{Title}
\usage{
linkpath(dataset, links)
}
\arguments{
\item{dataset}{name of dataset desired}

\item{links}{data frame with `shortname`, `address`, `longname`}
}
\value{
data frame with `shortname`, `address`, `longname`
}
\description{
Title
}
