% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met.R
\name{MetHarmony}
\alias{MetHarmony}
\title{Read Metabolite File for Harmonization}
\usage{
MetHarmony(dataset, links, annot, skiprows = 4, ...)
}
\arguments{
\item{dataset}{name of dataset}

\item{links}{dataframe of links to data files}

\item{annot}{annotation data frame}

\item{skiprows}{number of rows to skip}

\item{...}{ignore}
}
\value{
data frame with columns for strain, sex, animal, condition, trait, value
}
\description{
There are  minor differences among metabolite files.
To harmonize, we assume the following:
   Some number of rows (3-4) at top of file are skipped
   Columns from “compound” to “data_type” precede mouse data
   Mouse ID always begins with FFF-nn with “FFF” = founder, “nn” = animal
   (the rest of Mouse ID seems to change from file to file)
}
