% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleHarmony.R
\name{moduleHarmony}
\alias{moduleHarmony}
\title{Harmonize module object}
\usage{
moduleHarmony(
  datasetname,
  links,
  object,
  responsename = names(object[[1]]),
  ...,
  datatraits = FALSE,
  traitparams = FALSE
)
}
\arguments{
\item{datasetname}{name to give to dataset (here for compatibility)}

\item{links}{data frame with links (not used yet)}

\item{object}{object of class traitModule}

\item{responsename}{name of response to extract}

\item{...}{additional parameters ignored}

\item{datatraits}{incorporate `dataset` into `trait` if `TRUE`}

\item{traitparams}{incorporate key `params` into `trait` if `TRUE`}
}
\value{
harmonize object
}
\description{
Harmonize module object
}
