% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalscores.R
\name{normalscores}
\alias{normalscores}
\title{Normal Scores to Value Column of Object}
\usage{
normalscores(object, standard = FALSE)
}
\arguments{
\item{object}{data frame}

\item{standard}{standardize to mean 0 and variance 1 if `TRUE`}
}
\value{
data frame
}
\description{
Normal Scores to Value Column of Object
}
