% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nqrank.R
\name{nqrank}
\alias{nqrank}
\title{Normal scores as qnorm of rank}
\usage{
nqrank(x, jitter = FALSE, standard = FALSE)
}
\arguments{
\item{x}{vector of measurements}

\item{jitter}{jitter values slightly if `TRUE`}

\item{standard}{standardize to mean 0, variance 1 if `TRUE`}
}
\value{
vector of normal scores
}
\description{
Copied from github.com/kbroman/broman package
}
\examples{
nqrank(1:10)
}
