% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMixedModule.R
\name{shinyMixedModuleOutput}
\alias{shinyMixedModuleOutput}
\alias{shinyMixedModule}
\title{Shiny Module Output for Modules of Contrasts}
\usage{
shinyMixedModuleOutput(id)

shinyMixedModule(
  id,
  panel_par,
  main_par,
  traitContrast,
  mixedModule,
  customSettings = NULL
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitContrast}{reactive data frames}

\item{mixedModule}{static data frames}

\item{customSettings}{list of custom settings}
}
\value{
nothing returned

reactive object
}
\description{
Shiny Module Output for Modules of Contrasts

Shiny Module Server for Modules of Contrasts
}
